backend/pyproject.toml
[project]
name = "ainki-backend"
version = "0.1.0"
description = "AInki Backend - Spaced Repetition Learning Platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.104.1",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.9",
    "neo4j>=5.14.1",
    "langfuse>=3.5.2",
    "openai>=1.109.1",
    "PyPDF2>=3.0.1",
    "torch",
    "torchvision",
    "mineru[core]"
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true